Rename dbghelp.h to dbghelp.hpp
This is to avoid confusion with system DbgHelp.h that this header includes by itself. Note, original author named it dbghelp.hpp
